4_semak/Hacker_networks/task_5/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
selenium==4.17.0
python-dotenv==0.21.0
